requirements.txt
streamlit==1.2.0
numpy==1.19.2
altair==4.1.0
pandas==1.1.3
scipy==1.5.2
scikit-learn==0.23.2
pillow==8.0.1
jpeg==9b